requirements.txt
africastalking==1.1.5
Babel==2.6.0
certifi==2019.3.9
chardet==3.0.4
contextlib2==0.5.5
Django==1.11.20
django-phonenumber-field==2.3.1
djangorestframework==3.9.2
idna==2.8
phonenumbers==8.10.8
pytz==2018.9
requests==2.21.0
schema==0.7.0
urllib3==1.24.1
